plant_disease_app/requirements.txt
# Plant Disease Detection System - Requirements
# Python 3.9+

# Core Dependencies
tensorflow>=2.13.0
keras>=2.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Machine Learning & Preprocessing
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Application
streamlit>=1.28.0
streamlit-option-menu>=0.3.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0

# Optional: For better performance
# tensorflow-gpu>=2.13.0  # Uncomment for GPU support

# Optional: For model serving
# python-dotenv>=1.0.0
# pydantic>=2.0.0

# Development (Optional)
# jupyter>=1.0.0
# jupyterlab>=4.0.0
# pytest>=7.4.0
# black>=23.0.0
# pylint>=2.17.0
